Added Tag model to store info about id.

+++ data/src/main/java/justbucket/videolib/data/model/VideoEntity.kt
@@ -23,7 +23,7 @@ data class VideoEntity(
         var id: Long? = null,
 
         @ColumnInfo
-        var sourceId: Long,
+        var sourceId: Int,
 
         @ColumnInfo(name = COLUMN_VIDEO_TITLE)
         var title: String,
